Add Mathematica analysis notebook
The code in the initialisation cells is the backing that allows direct
symbolic analysis of the conditions generated in the Python file
    generate_mathematica.py
Importantly, the Python implementation of the expansion defined in the
paper makes _no_ assumptions about the form of the driving fields used
except that there's no off-resonant driving.  The Mathematica (and
consequently the Python code that stringifies the Python output into
Mathematica-readable code) assumes that the drives are all sums of
exponential terms, just like the simplest solutions given in the paper.