Add generation code for figure data
`phase/make.py` regenerates all the ".dat" files in that directory (up
to some ~1e-7 relative tolerance).  That file contains QuTiP
representations of the schemes defined in the paper, so these can be
used to produce the fidelity data stored in `pure` and `thermal` (there
are descriptive headers in those files), but there is no
batteries-included generation script for those in this directory because
we used the Imperial College London high-performance cluster to
distribute these calculations.